+++ app/Models/StorefrontSetting.php
@@ -294,6 +294,7 @@ class StorefrontSetting extends Model
         'marketplace_announcement_text',
         'marketplace_quote_enabled',
         'marketplace_business_accounts_enabled',
+        'marketplace_business_strip_enabled',
         'marketplace_trust_strip_enabled',
         'marketplace_deals_enabled',
         'marketplace_categories_enabled',
@@ -429,6 +430,7 @@ class StorefrontSetting extends Model
         'marketplace_announcement_enabled' => 'boolean',
         'marketplace_quote_enabled' => 'boolean',
         'marketplace_business_accounts_enabled' => 'boolean',
+        'marketplace_business_strip_enabled' => 'boolean',
         'marketplace_trust_strip_enabled' => 'boolean',
         'marketplace_deals_enabled' => 'boolean',
         'marketplace_categories_enabled' => 'boolean',
@@ -567,6 +569,7 @@ class StorefrontSetting extends Model
             $setting->marketplace_announcement_enabled ??= true;
             $setting->marketplace_quote_enabled ??= true;
             $setting->marketplace_business_accounts_enabled ??= true;
+            $setting->marketplace_business_strip_enabled ??= false;
             $setting->marketplace_trust_strip_enabled ??= true;
             $setting->marketplace_deals_enabled ??= true;
             $setting->marketplace_categories_enabled ??= true;